Changes to caching approach, and i18n added
Changes to caching approach, so that the JSON response array is stored rather than the generated HTML.  This means that every output is sanitized - repeated processing for each output, but it is guaranteed that the HTML is fully sanitized with no outside interference.  Added basic i18n for the one message currently given.