Initial work to add a Calendar Shortcode
Initial work to add a full month calendar shortcode.

The commit allows a responsive month calendar to be shown, by default the current month but also with a month specified by a supplied date within a month.

At present it does not provide functionality to move to a different month, nor does it display the details of an event, though you can click an event to get to its ChurchSuite details page.

More features to be added soon!