CSS fixes and month display in calendar
CSS Fixes to correct potential confusion from style .cs-calendar in the gliphs, and other minor CSS updates to add more consistency in labelling.

IDs added to all Groups and Events so they can be individually located in the DOM if needed

Addition of month name at the start of each month in the calendar, and changes to remove the date for each day from the inner html div so that the content and the date can be separately spaced if needed.

Minor colour shade changes for the calendar cell boundaries.